XRDDEV-584 & 681 Access rights PR comment fixes
* separate access right service methods from ServiceService into AccessRightService class
* api schema change: include Subject in ServiceClient
* IdentifierRepository
* test fixes
