+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/repository/IdentifierRepository.java
@@ -1,0 +1,89 @@
+/*
+ *  The MIT License
+ *  Copyright (c) 2018 Estonian Information System Authority (RIA),
+ *  Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
+ *  Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
+ *
+ *  Permission is hereby granted, free of charge, to any person obtaining a copy
+ *  of this software and associated documentation files (the "Software"), to deal
+ *  in the Software without restriction, including without limitation the rights
+ *  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ *  copies of the Software, and to permit persons to whom the Software is
+ *  furnished to do so, subject to the following conditions:
+ *
+ *  The above copyright notice and this permission notice shall be included in
+ *  all copies or substantial portions of the Software.
+ *
+ *  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ *  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ *  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ *  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ *  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ *  THE SOFTWARE.
+ */
+package org.niis.xroad.restapi.repository;
+
+import ee.ria.xroad.common.conf.serverconf.dao.IdentifierDAOImpl;
+import ee.ria.xroad.common.identifier.XRoadId;
+
+import lombok.extern.slf4j.Slf4j;
+import org.niis.xroad.restapi.util.PersistenceUtils;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+/**
+ * identifier repository
+ */
+@Slf4j
+@Repository
+@Transactional
+public class IdentifierRepository {
+
+    private final PersistenceUtils persistenceUtils;
+
+    @Autowired
+    public IdentifierRepository(PersistenceUtils persistenceUtils) {
+        this.persistenceUtils = persistenceUtils;
+    }
+
+    /**
+     * Executes a Hibernate saveOrUpdate(identifier)
+     * @param identifier
+     */
+    public void saveOrUpdate(XRoadId identifier) {
+        saveOrUpdate(identifier, false);
+    }
+
+    /**
+     * Executes a Hibernate saveOrUpdate(identifier) and flushes whole entityManager
+     * @param identifier
+     */
+    public void saveOrUpdateAndFlush(XRoadId identifier) {
+        saveOrUpdate(identifier, true);
+    }
+
+    /**
+     * Executes a Hibernate saveOrUpdate(identifier) and flushes whole entityManager
+     * @param identifier
+     * @param flush
+     */
+    public void saveOrUpdate(XRoadId identifier, boolean flush) {
+        persistenceUtils.getCurrentSession().saveOrUpdate(identifier);
+        if (flush) {
+            persistenceUtils.flush();
+        }
+    }
+
+    /**
+     * return all identifiers
+     */
+    public List<XRoadId> getIdentifiers() {
+        IdentifierDAOImpl identifierDAO = new IdentifierDAOImpl();
+        return identifierDAO.findAll(persistenceUtils.getCurrentSession(), XRoadId.class);
+    }
+}
+
