+++ src/proxy-ui-api/src/main/resources/openapi-definition.yaml
@@ -4008,35 +4008,12 @@ components:
       type: object
       description: service client
       required:
-        - id
-        - subject_type
-        - name
+        - subject
+        - rights_given_at
         - access_rights
       properties:
-        id:
-          type: string
-          format: text
-          description: a subsystem id <instance_id>:<member_class>:<member_code>:<subsystem> | globalgroup id <instance_id>:<group_code> | localgroup resource id in number format <id>
-          example: DEV:ORG:1234:Subsystem | DEV:security-server-owners | 123
-          minLength: 1
-          maxLength: 1279
-        subject_type:
-          $ref: '#/components/schemas/SubjectType'
-        name:
-          type: string
-          format: text
-          description: service client name
-          example: Security server owners
-          minLength: 1
-          maxLength: 255
-        local_group_code:
-          type: string
-          format: text
-          description: group code in case the object is a local group
-          example: My own Local group code
-          minLength: 1
-          maxLength: 255
-          readOnly: true
+        subject:
+          $ref: '#/components/schemas/Subject'
         rights_given_at:
           type: string
           format: date-time
